Change connect_by_number() so that SERVICE_SERVER sockets are always bound to IPv4 even on an IPv6-enabled client.  This fixes /detach and wserv when compiling with --enable-ipv6.
This does also mean that the internal identd, if used, won't listen on IPv6
either.  This isn't a problem yet, since the internal identd is used only on
WINNT and we don't support IPv6 there yet.

This will also need tweaking to support initiating DCC-over-IPv6.

Thanks to snadge and packet for working on this bug.


git-svn-id: svn://svn.code.sf.net/p/bitchx/code/trunk@162 13b04d17-f746-0410-82c6-800466cd88b0