Add back Makefile rules to create .config.h
This fixes a bug introduced in commit f2ae2327, which improved the Makefiles for parallel builds
but in doing so removed the rule to create an empty .config.h file.

.config.h isn't included in the source distribution - it's created by bxconfigure if you run it.
If you don't run it, the build systems needs to create an empty one.  This commit adds a rule to
both the toplevel Makefile.in and source/Makefile.in (the latter in case you're running 'make' just
in the source/ directory).  The toplevel one is required for building the modules tree under dll/,
since the module Makefiles don't have the right dependencies wired up.

This also makes 'distclean' remove the .config.h file.